Correctly choose nodes from a hash
The old technique (using the least significant 5 bits) gave equal chance
of a key being in level 1 or level 20 of a tree.  Those levels have
vastly different sizes so this caused many collisions.

The new approach assigns exponentially more elements to larger levels.
This is the same as if one numbered all nodes in the binary tree in
level order traversal.