Optimize deep trees with laziness
SHA1 generates trees that are 53-54 levels deep.  Real world trees need
only a few of those levels.  As we descend the tree, if we encounter a
branch that is not yet occupied at all, store our key-value pair as an
attribute on that variable.  If another key-value pair ends up passing
through this node later, we can push ourselves down one more level at
that point.

For a sample map with 9 keys, this code gives us a tree with 2 levels.
The old code gave one with 54 levels.

This improves performance by an order of magnitude in benchmarks.