Describe pmap's design
When returning to this code after 18 months, I had almost completely
forgotten how it worked.  The new Design comment at the top is intended
to give readers enough information about the data structure's design
that they can understand the code which implements it.

I also took this opportunity to use consistent nomeclature which I think
does a better job of describing the content of each node in the trie.