pmap: declare plump nodes where they're known
There were a few places in the code where a plump node was required but
we only checked that fact very late in the clause.  By declaring it
early in the clause, the clause can fail faster.

The declarations also help readers to understand what kind of node is
expected in each clause.