2x faster hashing
The old code used variant_sha1/2 for hashing keys.  The new code uses
sha_hash/3 which produces a list of bytes.  Summing the bytes is faster
than summing the hex digits.

This is still 100x slower than hashing with term_hash/2 but that doesn't
produce a large enough hash to be useful.