pmap: factor out tree parameters
I want to experiment with trees with different branching factors.  I'm
guessing that wider trees will perform better in certain scenarios.
This makes it easier to change them.

The performance cost of invoking these new predicates is minimal.  If it
becomes a problem, they can be replaced with goal_expansion so that they
have 0 cost.