pmap: more direct differ_in_one_child/5
With macro expansion, we can write differ_in_one_child/5 in such a way
that it performs most of its work with unification.  It also avoids the
need for the auxiliary predicate differ_in_one_child_

This improves runtime of t/pmap/huge.pl by 3.6%, reduces the number of
predicates and reduces the total amount of code.