map: remove dead or repetitive code
The predicates attr/2, hash_depth/2 and node/3 weren't used anymore.

The changes to kv were focused on removing the repetitive code between
kv/3 and the attribute unify hook.  The key observation is that a kv/3
declaration is the same as unifying an existing map with another map
that has only a single key-value pair.  Based on that observation, kv/3
can reuse attribute unification to perform the work.